add length check of auth_pub_key
