+++ src/lib/utils.ts
@@ -1,12 +1,14 @@
 import { type ClassValue, clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
-import bs58check from 'bs58check';
+import { isValidPoolAuthorityPubkey } from '@sv2-ui/shared';
 export {
   getBitcoinAddressError,
   getBitcoinAddressPlaceholder,
   getIdentifierError,
   isTomlSafeIdentifier,
   isValidBitcoinAddress,
+  isValidPoolAuthorityPubkey,
+  stripWrappingQuotes,
 } from '@sv2-ui/shared';
 
 /**
@@ -86,36 +88,6 @@ export function formatNumber(num: number, decimals: number = 0): string {
 export function calculateSharesPerMinute(shares: number, uptimeSecs: number): number {
   if (uptimeSecs === 0) return 0;
   return (shares / uptimeSecs) * 60;
-}
-
-// Pubkeys in pool docs / Discord are almost always shown wrapped in quotes,
-// and users copy them with the quotes. Strip one matched pair, then trim.
-export function stripWrappingQuotes(v: string): string {
-  const trimmed = v.trim();
-  if (
-    trimmed.length >= 2 &&
-    ((trimmed.startsWith('"') && trimmed.endsWith('"')) ||
-      (trimmed.startsWith("'") && trimmed.endsWith("'")))
-  ) {
-    return trimmed.slice(1, -1);
-  }
-  return trimmed;
-}
-
-// The SRI Noise-protocol authority key is base58check-encoded, so a corrupted
-// character or missing byte fails the checksum — which is exactly what
-// copy-paste mistakes produce. Wrapping quotes (the dominant mistake from
-// docs/Discord) are stripped here so the caller does not have to remember to
-// normalize before calling.
-export function isValidPoolAuthorityPubkey(v: string): boolean {
-  const normalized = stripWrappingQuotes(v);
-  if (!normalized) return false;
-  try {
-    bs58check.decode(normalized);
-    return true;
-  } catch {
-    return false;
-  }
 }
 
 export function getPoolAuthorityPubkeyError(v: string): string | null {
